requirements.txt
streamlit
psycopg2-binary
pyyaml